Group by search

+++ backend/src/main/java/com/devsuperior/dsvendas/dto/SaleDTO.java
@@ -9,12 +9,12 @@ public class SaleDTO {
 	private Long id;
 	private Integer visited;
 	private Integer deals;
-	private Integer amount;
+	private Double amount;
 	private LocalDate date;
 	
 	private SellerDTO seller;
 
-	public SaleDTO(Long id, Integer visited, Integer deals, Integer amount, LocalDate date, SellerDTO seller) {
+	public SaleDTO(Long id, Integer visited, Integer deals, Double amount, LocalDate date, SellerDTO seller) {
 		
 		this.id = id;
 		this.visited = visited;
@@ -50,10 +50,10 @@ public Integer getDeals() {
 public void setDeals(Integer deals) {
 	this.deals = deals;
 }
-public Integer getAmount() {
+public Double getAmount() {
 	return amount;
 }
-public void setAmount(Integer amount) {
+public void setAmount(Double amount) {
 	this.amount = amount;
 }
 public LocalDate getDate() {
